# code-repo-analysis/智能提醒助手/docs/api/reminder-service.openapi.yaml
openapi: 3.1.0
info:
  title: 智能提醒助手内部接口契约
  version: 1.0.0
  description: |
    本文件是《技术方案》4.3 章接口定义的字段级唯一事实源（IDL），覆盖三个内部 HTTP 接口、
    统一错误响应、事件信封与事件 payload、全部枚举取值。

    使用方式：
    - 开发：按本文件的字段名、类型、必填性实现 `internal/transport/http/dto`，字段只增不删不改名。
    - 测试：golden 用例直接取自本文件的 example，比对字段名与枚举值。

    边界：本文件只在本仓库内作为契约来源，不发布、不注册 PSM/BAM、不生成对外 SDK；
    本地入口仅绑定 127.0.0.1，由配置项 http.enabled 控制。文档与实现冲突时以本文件为准。
servers:
  - url: http://127.0.0.1:8080
    description: 本地回环入口；生产地址由接入方按既有服务规范决定
...
security:
  - staticToken: []
paths:
  /internal/v1/reminders/evaluate:
    post:
      operationId: evaluateReminder
      tags: [reminder]
      summary: 提醒资格判断
      description: |
        对单个用户在单个任务日期上执行一次提醒资格评估，覆盖 PRD FR-001、FR-002、FR-003、FR-004、FR-005。
        入口为每日分片评估、AppOpened 同步通道、任务状态变化事件和偏好变化事件。
        幂等键为 request_id + (user_id, task_date, strategy_version)；重复评估返回已有决策。
        业务拒绝（无任务、已完成、过期、关闭提醒、频控、无渠道）返回 200，用 status=REJECTED 与 reason_code 表达。
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/XCaller'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateReminderRequest'
            example:
              request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
              user_id: user-passive
              task_date: '2026-09-20'
              trigger: APP_OPENED
              last_open_at: '2026-09-15T11:20:00Z'
      responses:
        '200':
          description: 业务处理完成，含业务拒绝（status=REJECTED）
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateReminderResponse'
              examples:
                eligible:
                  summary: 被动型用户未完成任务，生成 Push 排程
                  value:
                    decision_id: decision-2291d3c8648abab22d033c693f0a3076
                    status: CREATED
                    reason_code: ELIGIBLE
                    should_remind: true
                    scheduled_at: '2026-09-20T19:30:00Z'
                    channel: push
                    user_segment: PASSIVE
                    strategy_version: v1
                returning_first_day:
                  summary: 回流首日，当天不发送提醒、不生成 Push 排程
                  value:
                    decision_id: decision-2291d3c8648abab22d033c693f0a3076
                    status: CREATED
                    reason_code: RETURNING_FIRST_DAY
                    should_remind: false
                    user_segment: RETURNING
                    strategy_version: v1
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/DependencyTimeout'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '504':
          $ref: '#/components/responses/DependencyTimeout'
  /internal/v1/reminders/dispatch-due:
    post:
      operationId: dispatchDue
      tags: [reminder]
      summary: 到期排程发送
      description: |
        扫描 status=SCHEDULED 且 scheduled_at<=now 的排程，抢占后二次校验任务与设置，再调用 Push 渠道（PRD FR-004、FR-005、FR-008）。
        不接受调用方传入执行时间或用户清单：执行时间由服务端 Clock 决定，处理对象由服务端按到期条件扫描。
        幂等：排程租约 + idempotency_key + 发送记录唯一键；单条排程最多 2 次发送尝试，重试复用同一幂等键。
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/XCaller'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchDueRequest'
            example:
              batch_size: 50
      responses:
        '200':
          description: 批次执行结果，只返回计数与游标，不返回用户级明细
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchDueResponse'
              example:
                scanned: 50
                claimed: 48
                sent: 41
                cancelled: 5
                failed: 1
                unknown: 1
                retried: 1
                finished: true
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/DependencyTimeout'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '504':
          $ref: '#/components/responses/DependencyTimeout'
  /internal/v1/reminders/{decision_id}/snooze:
    post:
      operationId: snoozeReminder
      tags: [reminder]
      summary: 一小时后再次提醒
      description: |
        用户点击“稍后提醒”后由 App 经既有网关同步调用（PRD FR-007）。
        服务端重新校验任务完成、过期、免打扰、截止时间与频控，通过则创建 schedule_type=SNOOZE 的 Push 排程。
        SNOOZE 排程独立计量，不占用当日系统主动 Push 上限（当天最多 2 次 Push）；延后提醒不再提供二次入口。
        幂等键为 request_id + decision_id；同一 request_id 重复调用返回 200 + status=REUSED 并复用原 schedule_id；
        对同一 decision_id 使用另一个 request_id 再次 snooze 属非法调用，返回 400 INVALID_REQUEST。
        decision_id 不存在、或存在但不属于请求中的 user_id 时，一律返回 200 + status=REJECTED + reason_code=DECISION_NOT_FOUND，
        两种情况响应体完全一致，不泄露资源存在性；403 只用于鉴权层，404 只用于未知路由。
      parameters:
        - $ref: '#/components/parameters/DecisionIDPath'
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/XCaller'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnoozeReminderRequest'
            example:
              request_id: 3c7d5e10-9a84-4f2e-8b71-2d5f6a7c8b9d
              user_id: user-passive
      responses:
        '200':
          description: 已创建延后排程（SCHEDULED）、命中幂等（REUSED）或业务拒绝（REJECTED）
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnoozeReminderResponse'
              examples:
                scheduled:
                  summary: 延后到 1 小时后
                  value:
                    schedule_id: schedule-6a1f0c93b7d24e58a0c5f8b1d3e7a942
                    status: SCHEDULED
                    reason_code: ELIGIBLE
                    snooze_until: '2026-09-20T20:30:00Z'
                    channel: push
                    schedule_type: SNOOZE
                rejected_by_quiet_hours:
                  summary: 进入免打扰且无可用窗口
                  value:
                    schedule_id: ''
                    status: REJECTED
                    reason_code: QUIET_HOURS
                rejected_decision_not_found:
                  summary: 决策不存在或不属于该用户，两种情况返回体一致
                  value:
                    schedule_id: ''
                    status: REJECTED
                    reason_code: DECISION_NOT_FOUND
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '408':
          $ref: '#/components/responses/DependencyTimeout'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/DependencyTimeout'
components:
  securitySchemes:
    staticToken:
      type: http
      scheme: bearer
      description: |
        本地为静态令牌白名单（Authenticator 端口），入口只绑定回环地址；
        生产替换为服务身份或 mTLS，并校验调用方身份与 user_id 归属。
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
      description: 调用方链路 ID；缺省时服务端生成，响应头回写
    XCaller:
      name: X-Caller
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
      description: 调用方标识，用于审计、限流维度与错误归因，如 reminder_scheduler、app-gateway
    DecisionIDPath:
      name: decision_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
      description: 原提醒决策 ID；必须属于请求中的 user_id，不属于时按业务拒绝处理，不泄露资源存在性
  headers:
    XRequestID:
      schema:
        type: string
      description: 请求追踪 ID，回写调用方传入值或服务端生成值
  responses:
    InvalidRequest:
      description: 参数缺失或前后矛盾（如 trigger 与 last_open_at 不匹配、未知枚举值）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: last_open_at must be empty when trigger is not APP_OPENED
            retryable: false
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    Unauthorized:
      description: 令牌缺失或无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: missing or invalid credential
            retryable: false
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    Forbidden:
      description: |
        鉴权层拒绝：调用方无权调用该动作，或无权操作该 user_id，与具体 decision_id 无关。
        decision_id 不存在或不属于该 user_id 不走本响应，按业务拒绝返回 200 + REJECTED + DECISION_NOT_FOUND。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: caller is not allowed to operate this user
            retryable: false
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    NotFound:
      description: |
        未知路由：路径本身不存在，未匹配到任何接口，不进入业务逻辑。
        decision_id 不存在或不属于该 user_id 不使用 404，按业务拒绝返回 200 + REJECTED + DECISION_NOT_FOUND。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: resource not found
            retryable: false
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    DependencyTimeout:
      description: 依赖超时，本次失败关闭，不发送旧快照
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: DEPENDENCY_TIMEOUT
            message: task service timeout
            retryable: true
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    RateLimited:
      description: 触发限流
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: rate limit exceeded
            retryable: true
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    InternalError:
      description: 未预期错误，不返回堆栈或敏感信息
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: DEPENDENCY_UNAVAILABLE
            message: internal error
            retryable: true
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
    FeatureDisabled:
      description: 服务级灰度开关关闭；这是不可重试的配置状态
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FEATURE_DISABLED
            message: feature is disabled
            retryable: false
            request_id: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
  schemas:
    EvaluateReminderRequest:
      type: object
      additionalProperties: true
      required: [request_id, user_id, task_date, trigger]
      description: 提醒资格评估请求（EvaluateReminder）
      properties:
        request_id:
          type: string
          minLength: 1
          maxLength: 64
          description: 调用请求唯一 ID；幂等键之一，重试必须复用同一取值
          example: 8f2c1a9e-6b41-4c2b-9f0d-7a1b2c3d4e5f
        user_id:
          type: string
          minLength: 1
          maxLength: 64
          description: 匿名用户 ID；不接受手机号、学号等可识别信息
          example: user-passive
        task_date:
          type: string
          format: date
          description: 用户本地任务日期（YYYY-MM-DD），不是服务器日期
          example: '2026-09-20'
        trigger:
          $ref: '#/components/schemas/Trigger'
        last_open_at:
          type: string
          description: |
            本次打开之前的最近一次打开时间（RFC3339 UTC）。
            仅 trigger=APP_OPENED 必填；无历史打开记录传空串 ""，不使用 null。
            服务端据此计算 days_since_last_open，不读取已被本次打开覆盖的记录。
          example: '2026-09-15T11:20:00Z'
    EvaluateReminderResponse:
      type: object
      required: [decision_id, status, reason_code, should_remind, strategy_version]
      description: 提醒资格评估响应；业务拒绝用 status=REJECTED 表达，仍返回 HTTP 200
      properties:
        decision_id:
          type: string
          description: 决策 ID；幂等命中时返回已有 ID
          example: decision-2291d3c8648abab22d033c693f0a3076
        status:
          type: string
          enum: [CREATED, REUSED, REJECTED]
          description: CREATED 新建决策；REUSED 幂等命中已有决策；REJECTED 业务拒绝
        reason_code:
          $ref: '#/components/schemas/ReasonCode'
        should_remind:
          type: boolean
          description: 是否进入提醒策略；status=REJECTED 时固定 false。服务端决策结果，不接受客户端传入
          example: true
        scheduled_at:
          type: string
          format: date-time
          description: 计划提醒时间（UTC RFC3339）；仅 should_remind=true 时返回
          example: '2026-09-20T19:30:00Z'
        channel:
          $ref: '#/components/schemas/Channel'
        user_segment:
          $ref: '#/components/schemas/UserSegment'
        strategy_version:
          type: string
          description: 本次决策使用的策略版本，取自服务端配置，不接受客户端传入
          example: v1
    DispatchDueRequest:
      type: object
      additionalProperties: true
      description: 到期发送请求；不接受执行时间与用户清单，由服务端 Clock 与到期条件决定
      properties:
        batch_size:
          type: integer
          format: int32
          minimum: 1
          maximum: 50
          default: 50
          description: 单次最多处理的到期排程数，上限受配置 dispatch_batch_size 约束
          example: 50
        cursor:
          type: string
          description: 分页游标，取上次响应返回的 next_cursor；首次调用省略，用于续跑积压
    DispatchDueResponse:
      type: object
      required: [scanned, claimed, sent, cancelled, failed, unknown, retried, finished]
      description: 到期发送批次结果；不返回 user_id、任务内容、推送文案等明细
      properties:
        scanned:
          type: integer
          description: 本次扫描到的到期排程数
        claimed:
          type: integer
          description: 成功抢占（SCHEDULED -> CLAIMED）的排程数
        sent:
          type: integer
          description: 渠道返回明确成功的数量
        cancelled:
          type: integer
          description: 发送前二次校验取消的数量（任务完成、过期、关闭提醒、免打扰、频控）
        failed:
          type: integer
          description: 明确失败且重试耗尽的数量
        unknown:
          type: integer
          description: 渠道状态未知、等待 reconcile 收敛的数量
        retried:
          type: integer
          description: 触发的重试次数；每条排程最多 1 次（max_send_attempts=2），复用同一 idempotency_key
        next_cursor:
          type: string
          description: 仍有未处理的到期排程时返回；为空表示本次已处理完
        finished:
          type: boolean
          description: 是否已无待处理排程
    SnoozeReminderRequest:
      type: object
      additionalProperties: true
      required: [request_id, user_id]
      description: 延后提醒请求（SnoozeReminder）；decision_id 在路径中
      properties:
        request_id:
          type: string
          minLength: 1
          maxLength: 64
          description: 幂等键；用户重复点击或调用方重试必须复用同一取值
          example: 3c7d5e10-9a84-4f2e-8b71-2d5f6a7c8b9d
        user_id:
          type: string
          minLength: 1
          maxLength: 64
          description: 匿名用户 ID；需与鉴权上下文一致，并与 decision_id 归属校验
          example: user-passive
    SnoozeReminderResponse:
      type: object
      required: [schedule_id, status, reason_code]
      properties:
        schedule_id:
          type: string
          description: 新建或已存在的 SNOOZE 排程 ID；status=REJECTED 时为空串
          example: schedule-6a1f0c93b7d24e58a0c5f8b1d3e7a942
        status:
          type: string
          enum: [SCHEDULED, REUSED, REJECTED]
          description: SCHEDULED 已创建延后排程；REUSED 幂等命中已有排程；REJECTED 未创建
        reason_code:
          $ref: '#/components/schemas/ReasonCode'
        snooze_until:
          type: string
          format: date-time
          description: 延后到的计划时间（UTC RFC3339）；仅 status=SCHEDULED/REUSED 时返回
          example: '2026-09-20T20:30:00Z'
        channel:
          $ref: '#/components/schemas/Channel'
        schedule_type:
          type: string
          enum: [INITIAL, SNOOZE]
          description: 排程类型；SNOOZE 表示用户主动延后，独立计量、不占当日系统主动 Push 上限
    ErrorResponse:
      type: object
      required: [code, message, retryable, request_id]
      description: 三个接口共用的错误响应
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 不含敏感信息的可读描述
        retryable:
          type: boolean
          description: 调用方是否可以安全重试；重试必须复用同一 request_id
        request_id:
          type: string
          description: 请求追踪 ID，与响应头 X-Request-ID 一致
    Trigger:
      type: string
      enum: [DAILY_BATCH, APP_OPENED, TASK_CHANGED, PREFERENCE_CHANGED]
      description: |
        评估触发来源：
        DAILY_BATCH 每日分片评估；APP_OPENED App 打开同步通道；TASK_CHANGED 任务状态变化异步事件；PREFERENCE_CHANGED 偏好变化异步事件。
        未知取值返回 400 INVALID_REQUEST。
    Channel:
      type: string
      enum: [push, in_app]
      description: 提醒渠道。本期只产生 push；in_app 为 P1 预留，本期不返回、不写入
    UserSegment:
      type: string
      enum: [PASSIVE, SELF_DRIVEN, INACTIVE, RETURNING]
      description: |
        本次分层结果，仅用于审计与效果分析，不作为跨日判定依据：
        PASSIVE 被动型；SELF_DRIVEN 自驱型；INACTIVE 流失型（连续 3 天以上未打开）；RETURNING 回流型（回流首日）。
    ReasonCode:
      type: string
      enum:
        - ELIGIBLE
        - NO_TASK
        - TASK_COMPLETED
        - TASK_EXPIRED
        - REMINDER_DISABLED
        - DAILY_CAP_REACHED
        - QUIET_HOURS
        - INSUFFICIENT_TIME
        - NO_AVAILABLE_CHANNEL
        - USER_INACTIVE
        - INSUFFICIENT_HISTORY
        - RETURNING_FIRST_DAY
        - DEFAULT_WINDOW
        - BEHAVIOR_DATA_DEGRADED
        - DECISION_NOT_FOUND
      description: |
        决策原因码，与 PRD 10.4 及技术方案 4.3.4 一致：
        ELIGIBLE 满足提醒条件；NO_TASK 当天没有任务；TASK_COMPLETED 当天任务已完成；TASK_EXPIRED 任务已过期；
        REMINDER_DISABLED 用户关闭提醒；DAILY_CAP_REACHED 达到每日系统主动 Push 上限；QUIET_HOURS 当前处于免打扰；
        INSUFFICIENT_TIME 剩余时间不足；NO_AVAILABLE_CHANNEL 没有可用渠道（本期为 Push 未授权或发送失败）；
        USER_INACTIVE 流失型不进入主动 Push 排程；INSUFFICIENT_HISTORY 历史行为不足 3 天，按被动型处理；
        RETURNING_FIRST_DAY 回流首日不发送提醒、不生成排程；DEFAULT_WINDOW 缺少有效历史，使用默认活跃窗口；
        BEHAVIOR_DATA_DEGRADED 行为数据降级，未据不完整数据命中回流或提醒归因规则。

        DECISION_NOT_FOUND 是接口层拒绝原因码，只由 snooze 的资源不可用路径产生，不代表评估链路的业务结论：
        decision_id 不存在、或存在但不属于请求中的 user_id 时均返回该值，两种情况响应体完全一致，调用方无法区分，
        避免用状态码差异枚举 decision_id。该取值已于 2026-09-21 经确认并收录进 PRD 10.4（与业务原因码分列说明）。
    ErrorCode:
      type: string
      enum:
        - REMINDER_ALREADY_EXISTS
        - DAILY_CAP_REACHED
        - TASK_COMPLETED
        - TASK_EXPIRED
        - REMINDER_DISABLED
        - QUIET_HOURS
        - NO_AVAILABLE_CHANNEL
        - BEHAVIOR_DATA_DEGRADED
        - DEPENDENCY_TIMEOUT
        - DEPENDENCY_UNAVAILABLE
        - FEATURE_DISABLED
        - CHANNEL_SEND_FAILED
        - CHANNEL_STATUS_UNKNOWN
        - CONTENT_INVALID
        - INVALID_REQUEST
      description: |
        错误码与渠道 error_code 的全部取值，与 PRD 10.4 及技术方案 4.3.4 一致：
        REMINDER_ALREADY_EXISTS 正常幂等命中；DAILY_CAP_REACHED/TASK_COMPLETED/TASK_EXPIRED/REMINDER_DISABLED/NO_AVAILABLE_CHANNEL 为正常业务拒绝；
        QUIET_HOURS 正常延后；BEHAVIOR_DATA_DEGRADED 正常降级；DEPENDENCY_TIMEOUT/DEPENDENCY_UNAVAILABLE 为可重试错误；
        FEATURE_DISABLED 表示服务级灰度开关关闭，返回 503 且不可重试；
        CHANNEL_SEND_FAILED 明确渠道错误，最多重试一次且幂等键不变；CHANNEL_STATUS_UNKNOWN 未知状态，先查询禁止盲目重发；INVALID_REQUEST 调用错误；
        CONTENT_INVALID 属内容校验错误（2026-09-21 FR-008 契约补齐），只写 reminder_delivery.error_code 与告警指标，不作为接口 ErrorResponse.code 返回，不重试。
...